+++ include/PPL/smoothed.h
